HSEARCH-2841 Move Elasticsearch Projection classes to their own file
