+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/query/impl/ElasticsearchHSQueryImpl.java
@@ -19,15 +19,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.regex.Pattern;
 
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.DoubleField;
-import org.apache.lucene.document.Field.Store;
-import org.apache.lucene.document.FloatField;
-import org.apache.lucene.document.IntField;
-import org.apache.lucene.document.LongField;
-import org.apache.lucene.document.StringField;
 import org.apache.lucene.search.Explanation;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.SortField;
@@ -102,8 +94,6 @@ public class ElasticsearchHSQueryImpl extends AbstractHSQuery {
 	private static final JsonParser JSON_PARSER = new JsonParser();
 
 	private static final Log LOG = LoggerFactory.make( Log.class );
-
-	private static final Pattern DOT = Pattern.compile( "\\." );
 
 	private static final String SPATIAL_DISTANCE_FIELD = "_distance";
 
@@ -844,199 +834,6 @@ public class ElasticsearchHSQueryImpl extends AbstractHSQuery {
 			return new EntityInfoImpl( typeId, documentBuilder.getIdPropertyName(), (Serializable) id, projections );
 		}
 
-	}
-
-	private abstract static class FieldProjection {
-
-		/**
-		 * Returns the value of the projected field as retrieved from the ES result and converted using the corresponding
-		 * field bridge. In case this bridge is not a 2-way bridge, the unconverted value will be returned.
-		 */
-		public abstract Object convertHit(JsonObject hit, ConversionContext conversionContext);
-
-		protected final JsonElement extractFieldValue(JsonObject parent, String projectedField) {
-			String field = projectedField;
-
-			if ( FieldHelper.isEmbeddedField( projectedField ) ) {
-				String[] parts = DOT.split( projectedField );
-				field = parts[parts.length - 1];
-
-				for ( int i = 0; i < parts.length - 1; i++ ) {
-					JsonElement newParent = parent.get( parts[i] );
-					if ( newParent == null ) {
-						return null;
-					}
-
-					parent = newParent.getAsJsonObject();
-				}
-			}
-
-			return parent.getAsJsonObject().get( field );
-		}
-
-	}
-
-	private static class TwoWayFieldBridgeProjection extends FieldProjection {
-
-		private final String absoluteName;
-		private final TwoWayFieldBridge bridge;
-		private final PrimitiveProjection defaultFieldProjection;
-		private final List<PrimitiveProjection> bridgeDefinedFieldsProjections;
-
-		public TwoWayFieldBridgeProjection(String absoluteName,
-				TwoWayFieldBridge bridge,
-				PrimitiveProjection defaultFieldProjection,
-				List<PrimitiveProjection> bridgeDefinedFieldsProjections) {
-			super();
-			this.absoluteName = absoluteName;
-			this.bridge = bridge;
-			this.defaultFieldProjection = defaultFieldProjection;
-			this.bridgeDefinedFieldsProjections = bridgeDefinedFieldsProjections;
-		}
-
-		@Override
-		public Object convertHit(JsonObject hit, ConversionContext conversionContext) {
-			return convertFieldValue( hit, conversionContext );
-		}
-
-		private Object convertFieldValue(JsonObject hit, ConversionContext conversionContext) {
-			Document tmp = new Document();
-
-			defaultFieldProjection.addDocumentField( tmp, hit, conversionContext );
-
-			// Add to the document the additional fields created when indexing the value
-			for ( PrimitiveProjection subProjection : bridgeDefinedFieldsProjections ) {
-				subProjection.addDocumentField( tmp, hit, conversionContext );
-			}
-
-			return conversionContext.twoWayConversionContext( bridge ).get( absoluteName, tmp );
-		}
-	}
-
-	private static class PrimitiveProjection extends FieldProjection {
-		private final TypeMetadata rootTypeMetadata;
-		private final String absoluteName;
-		private final ExtendedFieldType fieldType;
-
-		public PrimitiveProjection(TypeMetadata rootTypeMetadata, String absoluteName, ExtendedFieldType fieldType) {
-			super();
-			this.rootTypeMetadata = rootTypeMetadata;
-			this.absoluteName = absoluteName;
-			this.fieldType = fieldType;
-		}
-
-		public void addDocumentField(Document tmp, JsonObject hit, ConversionContext conversionContext) {
-			JsonElement jsonValue = extractFieldValue( hit.get( "_source" ).getAsJsonObject(), absoluteName );
-			if ( jsonValue == null || jsonValue.isJsonNull() ) {
-				return;
-			}
-			switch ( fieldType ) {
-				case INTEGER:
-					tmp.add( new IntField( absoluteName, jsonValue.getAsInt(), Store.NO ) );
-					break;
-				case LONG:
-					tmp.add( new LongField( absoluteName, jsonValue.getAsLong(), Store.NO ) );
-					break;
-				case FLOAT:
-					tmp.add( new FloatField( absoluteName, jsonValue.getAsFloat(), Store.NO ) );
-					break;
-				case DOUBLE:
-					tmp.add( new DoubleField( absoluteName, jsonValue.getAsDouble(), Store.NO ) );
-					break;
-				case UNKNOWN_NUMERIC:
-					throw LOG.unexpectedNumericEncodingType( rootTypeMetadata.getType(), absoluteName );
-				case BOOLEAN:
-					tmp.add( new StringField( absoluteName, String.valueOf( jsonValue.getAsBoolean() ), Store.NO ) );
-					break;
-				default:
-					tmp.add( new StringField( absoluteName, jsonValue.getAsString(), Store.NO ) );
-					break;
-			}
-		}
-
-		@Override
-		public Object convertHit(JsonObject hit, ConversionContext conversionContext) {
-			JsonElement jsonValue = extractFieldValue( hit.get( "_source" ).getAsJsonObject(), absoluteName );
-			if ( jsonValue == null || jsonValue.isJsonNull() ) {
-				return null;
-			}
-			switch ( fieldType ) {
-				case INTEGER:
-					return jsonValue.getAsInt();
-				case LONG:
-					return jsonValue.getAsLong();
-				case FLOAT:
-					return jsonValue.getAsFloat();
-				case DOUBLE:
-					return jsonValue.getAsDouble();
-				case UNKNOWN_NUMERIC:
-					throw LOG.unexpectedNumericEncodingType( rootTypeMetadata.getType(), absoluteName );
-				case BOOLEAN:
-					return jsonValue.getAsBoolean();
-				default:
-					return jsonValue.getAsString();
-			}
-		}
-	}
-
-	private static class JsonDrivenProjection extends FieldProjection {
-		private final String absoluteName;
-
-		public JsonDrivenProjection(String absoluteName) {
-			super();
-			this.absoluteName = absoluteName;
-		}
-
-		@Override
-		public Object convertHit(JsonObject hit, ConversionContext conversionContext) {
-			JsonElement value = extractFieldValue( hit.get( "_source" ).getAsJsonObject(), absoluteName );
-			if ( value == null || value.isJsonNull() ) {
-				return null;
-			}
-
-			// TODO: HSEARCH-2255 should we do it?
-			if ( !value.isJsonPrimitive() ) {
-				throw LOG.unsupportedProjectionOfNonJsonPrimitiveFields( value );
-			}
-
-			JsonPrimitive primitive = value.getAsJsonPrimitive();
-
-			if ( primitive.isBoolean() ) {
-				return primitive.getAsBoolean();
-			}
-			else if ( primitive.isNumber() ) {
-				// TODO HSEARCH-2255 this will expose a Gson-specific Number implementation; Can we somehow return an Integer,
-				// Long... etc. instead?
-				return primitive.getAsNumber();
-			}
-			else if ( primitive.isString() ) {
-				return primitive.getAsString();
-			}
-			else {
-				// TODO HSEARCH-2255 Better raise an exception?
-				return primitive.toString();
-			}
-		}
-	}
-	/**
-	 * A projection used whenever a given type has a one-way field bridge, which is forbidden.
-	 *
-	 * @author Yoann Rodiere
-	 */
-	private static class FailingOneWayFieldBridgeProjection extends FieldProjection {
-		private final String absoluteName;
-		private final Class<?> fieldBridgeClass;
-
-		public FailingOneWayFieldBridgeProjection(String absoluteName, Class<?> fieldBridgeClass) {
-			super();
-			this.absoluteName = absoluteName;
-			this.fieldBridgeClass = fieldBridgeClass;
-		}
-
-		@Override
-		public Object convertHit(JsonObject hit, ConversionContext conversionContext) {
-			throw LOG.projectingFieldWithoutTwoWayFieldBridge( absoluteName, fieldBridgeClass );
-		}
 	}
 
 	@Override
